fix(module: tree): add tests for testing dragging
