--- components/tree/nz-tree.service.spec.ts
@@ -1,3 +1,0 @@
-import { NzTreeService } from './nz-tree.service';
-
-export const nzTreeServiceStub: Partial<NzTreeService> = {};
